Made some refactoring and added custom nav controller.
