+++ Trippy/Trippy/App/AppDelegate.swift
@@ -9,13 +9,9 @@ import UIKit
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
-
-    private var appCoordinator: AppCoordinator?
     
     func application(_ application: UIApplication,
                      didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        appCoordinator = AppCoordinator(window: UIWindow())
-        
         return true
     }
 
@@ -25,14 +21,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                      configurationForConnecting connectingSceneSession: UISceneSession,
                      options: UIScene.ConnectionOptions) -> UISceneConfiguration {
         return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
-    }
-
-    func applicationDidEnterBackground(_ application: UIApplication) {
-        appCoordinator?.didEnterBackground()
-    }
-    
-    func applicationWillEnterForeground(_ application: UIApplication) {
-        appCoordinator?.willEnterForeground()
     }
 }
