+++ Trippy/Trippy/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
