+++ Trippy/Trippy/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
