+++ Trippy/Trippy/Resources/Assets.xcassets/Contents.json
